split symbol_manipulator to .h and .cpp
